Changed fields in ModuleDTO.
Moved DTO files to folder.
