+++ Server/Server.Tests/Class1.cs
@@ -7,7 +7,7 @@ namespace Server.Tests
     public class Class1
     {
         [Fact]
-        public void AlawysPassingTest()
+        public void AlwaysPassingTest()
         {
             Assert.True(true);
         }
