+++ Shared/SharedLibrary/Messages/EntityStateMessage.cs
@@ -1,5 +1,5 @@
 ﻿using System.Text.Json;
-using SharedLibrary;
+using SharedLibrary.DTOs.EntitiesDTO;
 
 namespace SharedLibrary.Messages
 {
